+++ Source/WebCore/platform/graphics/filters/FilterFunction.h
@@ -25,11 +25,8 @@
 
 #pragma once
 
-#include "FilterEffectGeometry.h"
 #include "FilterImage.h"
-#include "FilterImageVector.h"
 #include "FloatRect.h"
-#include "LengthBox.h"
 #include <wtf/RefCounted.h>
 #include <wtf/text/AtomString.h>
